Add Python APM to CSH's list of APMs
